[FOEPD-3760] harden grid-data adapter interface
Review pass on the interface-only branch. The 11 changes:

1. Replace "seam" with "interface" / "adapter pattern" in docstrings.

2. Remove ``SQLGridAdapter`` stub + its factory branch + its test —
   dead weight; the protocol-conformance check it provided is
   adequately covered by ``test_mongo_grid_adapter`` and the
   ``isinstance`` runtime check.

3. Keep ``@runtime_checkable`` on ``GridDataAdapter`` — the
   ``isinstance`` check in ``test_factory`` is cheap and harmless.

4. Replace tuple returns on ``paginate_samples`` and
   ``count_field_values`` with frozen dataclasses
   :class:`SamplesPage` (``samples`` / ``has_more``) and
   :class:`ValuePickerResult` (``total`` / ``page``). Forces named
   access at the call site (``samples.py``, ``routes/values.py``,
   ``test_mongo_grid_adapter``).

5. Move ``FIFTYONE_GRID_BACKEND`` selection to
   ``fiftyone.config.grid_backend`` (new ``parse_string`` entry in
   ``FiftyOneConfig`` reading the env var as default). Factory reads
   ``fo.config.grid_backend`` instead of ``os.environ`` directly. No
   import cycle: factory imports ``fiftyone`` at module top.

6. Convert interface + Mongo adapter + factory docstrings to
   Google-style (``Args:`` / ``Returns:`` / ``Raises:``).

7. Strip implementation / call-site notes from the interface
   docstrings. Contract first; implementation lives on the adapter.

8. Drop ``estimated_sample_count`` from the interface entirely.
   In FO, Mongo is the canonical source of truth for sample
   documents; the grid backend filters over them. The dataset's
   total sample count is a Mongo metadata fact regardless of grid
   backend, so the resolver in ``query.py`` reverts to a direct
   ``info.context.db[...].estimated_document_count()`` call —
   mirroring ``estimated_frame_count`` two lines below it. No
   extra Mongo lookup, no leaked Mongo collection name in the
   adapter interface.

9. Rewrite ``db/mongo.py`` module docstring; no more
   "today's state" framing.

10. Extract Mongo lightning pipeline helpers from
    ``fiftyone/server/lightning.py`` to a new private module
    ``fiftyone/server/db/_mongo_lightning.py``. Breaks the
    ``MongoGridAdapter`` ↔ ``lightning_resolver`` import cycle —
    the adapter now imports helpers at module top, and the
    resolver defers its ``get_grid_adapter`` lookup to request
    time. Exports renamed from ``_resolve_lightning_path_queries``
    / ``_do_async_pooled_queries`` to public-leading
    ``resolve_lightning_path_queries`` / ``do_async_pooled_queries``
    since they are now legitimate module-level exports.

11. Delete ``MetadataAdapter`` entirely
    (interface + ``MongoMetadataAdapter`` + ``get_metadata_adapter``
    factory + test). Revert ``dataloader.py`` and ``paginator.py``
    to direct Mongo calls. The metadata interface was always-Mongo
    by docstring fiat, didn't enable swap-ability, and missed
    other metadata collections (saved_views, …) anyway. Also fixed
    the pre-existing shared-config mutation bug in
    ``dataloader.py`` (use a local ``key`` variable instead of
    mutating ``config.key``).

Tests:

* 37/37 across ``tests/unittests/server/db/`` and the existing
  ``server_lightning_tests`` / ``server_aggregations_tests`` /
  ``server_samples_tests`` / ``server_view_tests``.
