+++ fiftyone/server/db/__init__.py
@@ -1,28 +1,20 @@
 """
-FiftyOne Server data adapter package.
+FiftyOne Server grid-data adapter package.
 
-Provides a seam between server resolvers and the underlying database. The
-``GridDataAdapter`` interface is the swappable surface (sample/scene reads
-that may eventually be served from a SQL/BigQuery backend); the
-``MetadataAdapter`` interface holds metadata reads that are always Mongo.
+Provides an adapter-pattern interface between server resolvers and the
+underlying database for grid / sidebar / sample reads. The
+``GridDataAdapter`` interface is the swappable surface; a non-Mongo
+implementation can plug in.
 
 | Copyright 2017-2026, Voxel51, Inc.
 | `voxel51.com <https://voxel51.com/>`_
 |
 """
 
-from fiftyone.server.db.factory import (
-    get_grid_adapter,
-    get_metadata_adapter,
-)
-from fiftyone.server.db.interface import (
-    GridDataAdapter,
-    MetadataAdapter,
-)
+from fiftyone.server.db.factory import get_grid_adapter
+from fiftyone.server.db.interface import GridDataAdapter
 
 __all__ = [
     "GridDataAdapter",
-    "MetadataAdapter",
     "get_grid_adapter",
-    "get_metadata_adapter",
 ]
